+++ R/cnetplot.R
@@ -1,64 +1,97 @@
 ##' @rdname cnetplot
 ##' @exportMethod cnetplot
 setMethod("cnetplot", signature(x = "enrichResult"),
-          function(x, showCategory = 5,
-                   foldChange = NULL, layout = "kk", ...) {
-              cnetplot.enrichResult(x, showCategory = showCategory,
-                                    foldChange = foldChange, layout = layout, ...)
+          function(x,  ...) {
+              cnetplot.enrichResult(x,  ...)
           })
 
 ##' @rdname cnetplot
 ##' @exportMethod cnetplot
 setMethod("cnetplot", signature(x = "list"),
-          function(x, showCategory = 5,
-                   foldChange = NULL, layout = "kk", ...) {
-              cnetplot.enrichResult(x, showCategory = showCategory,
-                                    foldChange = foldChange, layout = layout, ...)
+          function(x, ...) {
+              cnetplot.enrichResult(x, ...)
           })
 
 ##' @rdname cnetplot
 ##' @exportMethod cnetplot
 setMethod("cnetplot", signature(x = "gseaResult"),
-          function(x, showCategory = 5,
-                   foldChange = NULL, layout = "kk", ...) {
-              cnetplot.enrichResult(x, showCategory = showCategory,
-                                    foldChange = foldChange, layout = layout, ...)
+          function(x, ...) {
+              cnetplot.enrichResult(x, ...)
           })
 
 ##' @rdname cnetplot
 ##' @exportMethod cnetplot
 setMethod("cnetplot", signature(x = "compareClusterResult"),
-          function(x, showCategory = 5,
-                   foldChange = NULL, layout = "kk", ...) {
-              cnetplot.compareClusterResult(x, showCategory = showCategory,
-                                    foldChange = foldChange, layout = layout, ...)
+          function(x, ...) {
+              cnetplot.compareClusterResult(x, ...)
           })
 
 
 ##' @rdname cnetplot
 ##' @param colorEdge Logical, whether coloring edge by enriched terms, the default value is FALSE. 
+##' Will be removed in the next version.
 ##' @param circular Logical, whether using circular layout, the default value is FALSE.
+##' Will be removed in the next version.
 ##' @param node_label Select which labels to be displayed.
 ##' one of 'category', 'gene', 'all'(the default) and 'none'.
 ##' @param cex_category Number indicating the amount by which plotting category
 ##' nodes should be scaled relative to the default, the default value is 1.
+##' Will be removed in the next version.
 ##' @param cex_gene Number indicating the amount by which plotting gene nodes
 ##' should be scaled relative to the default, the default value is 1.
+##' Will be removed in the next version.
 ##' @param cex_label_category Scale of category node label size, the 
 ##' default value is 1.
+##' Will be removed in the next version.
 ##' @param cex_label_gene Scale of gene node label size, the default value is 1.
+##' Will be removed in the next version.
 ##' @param color_category Color of category node.
+##' Will be removed in the next version.
 ##' @param color_gene Color of gene node.
+##' Will be removed in the next version.
 ##' @param shadowtext select which node labels to use shadow font,
 ##' one of 'category', 'gene', 'all' and 'none', default is 'all'.
+
+##' @param color.params list, the parameters to control the attributes of highlighted nodes and edges.
+##' see the color.params in the following.
+##' color.params control the attributes of highlight, it can be referred to the following parameters:
+##'     \itemize{
+##'         \item \code{foldChange} Fold Change of nodes for enrichResult, or size of nodes for compareClusterResult, 
+##'         the default value is NULL. 
+##'         \item \code{edge} Logical, whether coloring edge by enriched terms, the default value is FALSE. 
+##'         \item \code{category} Color of category node.
+##'         \item \code{gene} Color of gene node.
+##'     }
+##' @param cex.params list, the parameters to control the size of nodes and lables.
+##' see the cex.params in the following.
+##' cex.params control the attributes of highlight, it can be referred to the following parameters:
+##'     \itemize{
+##'         \item \code{foldChange} only used in compareClusterResult object, fold Change of nodes, the default value is NULL. 
+##'         If the user provides the Fold Change value of the nodes, 
+##'         it can be used to set the size of the gene node.
+##'         \item \code{category_node} Number indicating the amount by which plotting category
+##'         nodes should be scaled relative to the default, the default value is 1.
+##'         \item \code{gene_node} Number indicating the amount by which plotting gene nodes
+##'         should be scaled relative to the default, the default value is 1.
+##'         \item \code{category_label} Scale of category node label size, the 
+##'         default value is 1.
+##'         \item \code{gene_label} Scale of gene node label size, the default value is 1.
+##'     }
+##' @param hilight.params list, the parameters to control the attributes of highlighted nodes and edges.
+##' see the hilight.params in the following.
+##' hilight.params control the attributes of highlight, it can be referred to the following parameters:
+##'     \itemize{
+##'         \item \code{category} category nodes to be highlight.
+##'         \item \code{alpha_hilight} alpha of highlighted nodes.
+##'         \item \code{alpha_no_hilight} alpha of unhighlighted nodes.
+##'     }
 ##' @importFrom ggraph geom_edge_arc
 ##' @importFrom ggplot2 scale_colour_gradient2
-##' @importFrom ggnewscale new_scale_color
-##' @importFrom ggnewscale new_scale
+##' @importFrom rlang enquo
 ##' @author Guangchuang Yu
 cnetplot.enrichResult <- function(x,
                      showCategory = 5,
-                     foldChange   = NULL,
+                     foldChange = NULL,
                      layout = "kk",
                      colorEdge = FALSE,
                      circular = FALSE,
@@ -70,11 +103,118 @@ cnetplot.enrichResult <- function(x,
                      color_category = "#E5C494",
                      color_gene = "#B3B3B3",
                      shadowtext = "all",
+                     color.params=list(
+                         foldChange = NULL,
+                         edge = FALSE,           
+                         category = "#E5C494",   
+                         gene = "#B3B3B3"        
+                    ),
+                     cex.params=list(
+                         category_node = 1,      
+                         gene_node = 1,          
+                         category_label = 1,     
+                         gene_label = 1         
+                    ),
+                     hilight.params=list(        
+                         category = NULL,      
+                         alpha_hilight = 1,      
+                         alpha_no_hilight = 0.3  
+                     ),
                      ...) {
 
     label_size_category <- 5
     label_size_gene <- 5
     node_label <- match.arg(node_label, c("category", "gene", "all", "none"))
+    
+    params_df <- as.data.frame(rbind(
+        c("foldChange", "color.params", "foldChange"),
+        c("colorEdge", "color.params", "edge"),
+        c("color_category", "color.params", "category"),
+        c("color_gene", "color.params", "gene"),
+
+        c("cex_category", "cex.params", "category_node"),
+        c("cex_gene", "cex.params", "gene_node"),
+        c("cex_label_category", "cex.params", "category_label"),
+        c("cex_label_gene", "cex.params", "gene_label"))
+    )
+    colnames(params_df) <- c("original", "listname", "present")
+    rownames(params_df) <- params_df$original
+
+
+    
+    default.color.params <- list(
+        foldChange = NULL,
+        edge = FALSE,           
+        category = "#E5C494",   
+        gene = "#B3B3B3"        
+    )
+    default.cex.params <- list(
+        category_node = 1,       
+        gene_node = 1,                
+        category_label = 1,      
+        gene_label = 1        
+    )
+    default.hilight.params <- list(
+        category = NULL,
+        alpha_hilight = 1,
+        alpha_no_hilight = 0.3
+    )
+
+    # use modifyList to change the values of parameter 
+    color.params <- modifyList(default.color.params, color.params)
+    cex.params <- modifyList(default.cex.params, cex.params)
+    hilight.params <- modifyList(default.hilight.params, hilight.params)
+    params_list <- list(x = x,
+            showCategory = showCategory,
+            foldChange = foldChange,
+            layout = layout,
+            colorEdge = colorEdge,
+            circular = circular,
+            node_label = node_label,
+            cex_category = cex_category,
+            cex_gene = cex_gene,
+            cex_label_category = cex_label_category,
+            cex_label_gene = cex_label_gene,
+            color_category = color_category,
+            color_gene = color_gene,
+            shadowtext = shadowtext,
+            color.params = color.params, 
+            cex.params = cex.params,
+            hilight.params = hilight.params
+    )
+
+    # get all parameters value
+    args <- as.list(match.call())
+    removed_params <- intersect(params_df$original, names(args))
+    if (length(removed_params) > 0) {
+        for (i in removed_params) {
+            params_list[[params_df[i, 2]]][[params_df[i, 3]]] <- get(i)
+            warn <- get_param_change_message(i, params_df)
+            warning(warn)
+        }
+    }
+
+    
+    color.params <- params_list[["color.params"]]
+    cex.params <- params_list[["cex.params"]]
+    hilight.params <- params_list[["hilight.params"]]
+ 
+    foldChange <- color.params[["foldChange"]]                               
+    colorEdge <- color.params[["edge"]]
+    color_category <- color.params[["category"]]
+    color_gene <- color.params[["gene"]]
+
+
+    cex_category <- cex.params[["category_node"]]
+    cex_gene <- cex.params[["gene_node"]]
+    cex_label_category <- cex.params[["category_label"]]
+    cex_label_gene <- cex.params[["gene_label"]]
+
+
+    hilight_category <- hilight.params[["category"]]
+    alpha_hilight <- hilight.params[["alpha_hilight"]]
+    alpha_nohilight <- hilight.params[["alpha_no_hilight"]]
+
     if (circular) {
         layout <- "linear"
         geom_edge <- geom_edge_arc
@@ -89,7 +229,7 @@ cnetplot.enrichResult <- function(x,
     if (shadowtext == "category") shadowtext_category <- TRUE
     if (shadowtext == "gene") shadowtext_gene <- TRUE
     geneSets <- extract_geneSets(x, showCategory)
-
+    
     g <- list2graph(geneSets)
 
     if (!inherits(x,  "list")) {
@@ -101,12 +241,33 @@ cnetplot.enrichResult <- function(x,
     n <- length(geneSets)
     V(g)$size[1:n] <- size
     node_scales <- c(rep(cex_category, n), rep(cex_gene, (length(V(g)) - n)))
+
+    # add edge alpha
+    hilight_category <- intersect(hilight_category, names(geneSets))
+
+    if (!is.null(hilight_category) && length(hilight_category) > 0) {
+        edges <- attr(E(g), "vnames")
+        E(g)$alpha <- rep(alpha_nohilight, length(E(g)))
+        hilight_edge <- grep(paste(hilight_category, collapse = "|"), edges)
+        hilight_gene <- edges[hilight_edge]
+        hilight_gene <- gsub(".*\\|", "", hilight_gene)
+        E(g)$alpha[hilight_edge] <- min(0.8, alpha_hilight)
+    } else {
+        E(g)$alpha <- rep(0.8, length(E(g)))
+    }
+
+    show_legend <- c(FALSE, TRUE)
+    names(show_legend) <- c("alpha", "color") 
     if (colorEdge) {
         E(g)$category <- rep(names(geneSets), sapply(geneSets, length))
-        edge_layer <- geom_edge(aes_(color = ~category), alpha=.8)
+        edge_layer <- geom_edge(aes_(color = ~category, alpha = ~I(alpha)), 
+            show.legend = show_legend)
     } else {
-        edge_layer <- geom_edge(alpha=.8, colour='darkgrey')
+        edge_layer <- geom_edge(aes_(alpha = ~I(alpha)), colour='darkgrey',
+            show.legend = FALSE)
     }
+
+
 
     if (!is.null(foldChange)) {
         fc <- foldChange[V(g)$name[(n+1):length(V(g))]]
@@ -117,42 +278,39 @@ cnetplot.enrichResult <- function(x,
         names(show_legend) <- c("color", "size")
         p <- ggraph(g, layout=layout, circular = circular)
         p$data[-(1:n), "size"] <- 3 * cex_gene
+ 
+        p <- node_add_alpha(p, hilight_category, hilight_gene, alpha_nohilight, alpha_hilight)
+       
+        alpha_category <- c(rep(1, n), rep(0, nrow(p$data)-n))
+        alpha_gene <- c(rep(0, n), rep(1, nrow(p$data)-n))
+
+        if (!is.null(hilight_category) && length(hilight_category) > 0) {
+            alpha_category <- c(rep(alpha_nohilight, n), rep(0, nrow(p$data)-n))
+            alpha_gene <- c(rep(0, n), rep(alpha_nohilight, nrow(p$data)-n))
+            alpha_gene[match(hilight_gene, p$data$name)] <- alpha_hilight
+            alpha_gene[match(hilight_category, p$data$name)] <- alpha_hilight
+        }
 
         p <- p + edge_layer +
             geom_node_point(aes_(size=~size), color=I(color_category),
                         data = NULL, show.legend = show_legend,
-                        alpha = c(rep(1, n), rep(0, nrow(p$data)-n))) +
+                        alpha = I(alpha_category)) +
             ggnewscale::new_scale_color() +
             geom_node_point(aes_(color=~as.numeric(as.character(color)), size=~size),
-                data = NULL, alpha = c(rep(0, n), rep(1, nrow(p$data)-n))) +
+                data = NULL, alpha = I(alpha_gene)) +
             scale_size(range=c(3, 8) * cex_category) +  
             scale_colour_gradient2(name = "fold change", low = "blue",
                                    mid = "white", high = "red",
                                    guide = guide_colorbar(order = 2))
 
-
-        # p <- p + edge_layer +
-        #     geom_node_point(aes_(color=~I(color_category), size=~size),
-        #         data = p$data[1:n, ]) +
-        #     scale_size(range=c(3, 8) * cex_category) +
-        #     ggnewscale::new_scale_color() +
-        #     geom_node_point(aes_(color=~as.numeric(as.character(color)), size=~I(3 * cex_gene)),
-        #                     data = p$data[-(1:n), ], show.legend = show_legend) +
-        #     scale_colour_gradient2(name = "fold change", low = "blue",
-        #                            mid = "white", high = "red",
-        #                            guide = guide_colorbar(order = 2))
     } else {
         V(g)$color <- color_gene
         V(g)$color[1:n] <- color_category
         p <- ggraph(g, layout=layout, circular=circular)
-        # p <- p + edge_layer +
-        #     geom_node_point(aes_(color=~I(color), size=~size), data = p$data[1:n, ]) +
-        #     scale_size(range=c(3, 8) * cex_category) +
-        #     geom_node_point(aes_(color=~I(color), size=~I(3 * cex_gene)),
-        #                     data = p$data[-(1:n), ], show.legend = FALSE) 
         p$data[-(1:n), "size"] <- 3 * cex_gene
+        p <- node_add_alpha(p, hilight_category, hilight_gene, alpha_nohilight, alpha_hilight)
         p <- p + edge_layer +
-            geom_node_point(aes_(color=~I(color), size=~size))+
+            geom_node_point(aes_(color=~I(color), size=~size, alpha=~I(alpha)))+
             scale_size(range=c(3, 8) * cex_category) 
 
     }
@@ -170,32 +328,38 @@ cnetplot.enrichResult <- function(x,
     } else if (node_label == "all") {
         p <- add_node_label(p = p, data = NULL,
             label_size_node = c(rep(label_size_category, n), rep(label_size_gene, nrow(p$data)-n)),
-            cex_label_node = c(rep(cex_label_category, n), rep(cex_label_gene,, nrow(p$data)-n)), 
+            cex_label_node = c(rep(cex_label_category, n), rep(cex_label_gene, nrow(p$data)-n)), 
             shadowtext = shadowtext_gene)
-        # p <- add_node_label(p = p, data = p$data[-c(1:n),], label_size_node = label_size_gene,
-        #     cex_label_node = cex_label_gene, shadowtext = shadowtext_gene)
-        # p <- add_node_label(p = p, data = p$data[1:n,], label_size_node = label_size_category,
-        #     cex_label_node = cex_label_category, shadowtext = shadowtext_category)
     }
     if (!is.null(foldChange)) {
         p <- p + guides(size  = guide_legend(order = 1), 
                         color = guide_colorbar(order = 2))
     }
-    return(p)
+    return(p + guides(alpha = "none"))
 }
 
 
 ##' @param split Separate result by 'category' variable.
 ##' @param pie Proportion of clusters in the pie chart, one of 'equal' (default) and 'Count'.
+##' Will be removed in the next version.
 ##' @param legend_n Number of circle in legend, the default value is 5.
+##' Will be removed in the next version.
 ##' @param x_loc,y_loc The location of scatterpie legend.
+##' Will be removed in the next version.
+##' @param pie.params list, the parameters to control the attributes of pie nodes.
+##' see the pie.params in the following.
+##' pie.params control the attributes of pie nodes, it can be referred to the following parameters:
+##'     \itemize{
+##'         \item \code{pie} proportion of clusters in the pie chart, one of 'equal' (default) and 'Count'.
+##'         \item \code{legend_n} number of circle in legend.
+##'         \item \code{legend_loc_x, legend_loc_y} The location of scatterpie legend.
+##'     }
 ##' @importFrom ggraph geom_edge_arc
 ##' @noRd
 cnetplot.compareClusterResult <- function(x,
                      showCategory = 5,
                      foldChange   = NULL,
                      layout = "kk",
-                     colorEdge = FALSE,
                      circular = FALSE,
                      node_label = "all",
                      split=NULL,
@@ -208,12 +372,132 @@ cnetplot.compareClusterResult <- function(x,
                      cex_label_category = 1,
                      cex_label_gene = 1,
                      shadowtext = "all",
+                     pie.params = list(
+                         pie = "equal",            
+                         legend_n = 5,             
+                         legend_loc_x = NULL,      
+                         legend_loc_y = NULL       
+                     ),
+                     cex.params=list(
+                         foldChange = NULL,
+                         category_node = 1,        
+                         gene_node = 1,            
+                         category_label = 1,       
+                         gene_label = 1           
+                    ),
+                     hilight.params=list(
+                         category = NULL,
+                         alpha_hilight = 1,
+                         alpha_no_hilight = 0.3
+                     ),  
                      ...) {
 
     label_size_category <- 2.5
     label_size_gene <- 2.5
     range_category_size <- c(3, 8)
     range_gene_size <- c(3, 3)
+
+    # change parameter name
+    ##############################################################
+    params_df <- as.data.frame(rbind(
+        c("pie", "pie.params", "pie"),
+        c("legend_n", "pie.params", "legend_n"),
+        c("x_loc", "pie.params", "legend_loc_x"),
+        c("y_loc", "pie.params", "legend_loc_y"),
+
+        c("foldChange", "cex.params", "foldChange"),
+        c("cex_category", "cex.params", "category_node"),
+        c("cex_gene", "cex.params", "gene_node"),
+        c("cex_label_category", "cex.params", "category_label"),
+        c("cex_label_gene", "cex.params", "gene_label"))
+    )
+    colnames(params_df) <- c("original", "listname", "present")
+    rownames(params_df) <- params_df$original
+
+    default.pie.params <- list(
+        pie = "equal",          
+        legend_n = 5,             
+        legend_loc_x = NULL,      
+        legend_loc_y = NULL           
+    )
+    default.cex.params <- list(
+        foldChange = NULL,     
+        category_node = 1,      
+        gene_node = 1,          
+        category_label = 1,     
+        gene_label = 1               
+    )
+    default.hilight.params <- list(
+        category = NULL,
+        alpha_hilight = 1,
+        alpha_no_hilight = 0.3
+    )
+
+    # use modifyList to change the values of parameter 
+    pie.params <- modifyList(default.pie.params, pie.params)
+    cex.params <- modifyList(default.cex.params, cex.params)
+    hilight.params <- modifyList(default.hilight.params, hilight.params)
+
+    params_list <- list(x = x,
+        showCategory = showCategory,
+        foldChange = foldChange,
+        layout = layout,
+        circular = circular,
+        node_label = node_label ,
+        split = split,
+        pie = pie,
+        cex_category = cex_category,
+        cex_gene = cex_gene,
+        legend_n = legend_n,
+        x_loc = x_loc,
+        y_loc = y_loc,
+        cex_label_category = cex_label_category,
+        cex_label_gene = cex_label_gene,
+        shadowtext = shadowtext,
+        pie.params = pie.params,
+        cex.params = cex.params,
+        hilight.params = hilight.params  
+    )
+
+    # get all parameters value
+    args <- as.list(match.call())
+    removed_params <- intersect(params_df$original, names(args))
+    if (length(removed_params) > 0) {
+        for (i in removed_params) {
+            params_list[[params_df[i, 2]]][[params_df[i, 3]]] <- get(i)
+            warn <- get_param_change_message(i, params_df)
+            warning(warn)
+        }
+    }
+
+    pie.params <- params_list[["pie.params"]]
+    cex.params <- params_list[["cex.params"]]
+    hilight.params <- params_list[["hilight.params"]]
+
+    pie <- pie.params[["pie"]]
+    legend_n <- pie.params[["legend_n"]]
+    x_loc <- pie.params[["legend_loc_x"]]
+    y_loc <- pie.params[["legend_loc_y"]]
+
+
+    foldChange <- cex.params[["foldChange"]]
+    cex_category <- cex.params[["category_node"]]
+    cex_gene <- cex.params[["gene_node"]]
+    cex_label_category <- cex.params[["category_label"]]
+    cex_label_gene <- cex.params[["gene_label"]]
+
+
+    # default.hilight.params <- list(
+    #     category = NULL,
+    #     alpha_hilight = 1,
+    #     alpha_no_hilight = 0.3
+    # )
+    # hilight.params <- reset_params(defaultp=default.hilight.params, 
+    #                                inputp=enquo(hilight.params))
+    hilight_category <- hilight.params[["category"]]
+    alpha_hilight <- hilight.params[["alpha_hilight"]]
+    alpha_nohilight <- hilight.params[["alpha_no_hilight"]]
+
     if (is.logical(shadowtext)) {
         shadowtext <- ifelse(shadowtext, "all", "none")
     }
@@ -233,6 +517,7 @@ cnetplot.compareClusterResult <- function(x,
     }
     ## Data structure transformation, combining the same ID (Description) genes
     y_union <- merge_compareClusterResult(y)
+    hilight_category <- intersect(hilight_category, y$Description)
     node_label <- match.arg(node_label, c("category", "gene", "all", "none"))
     if (circular) {
         layout <- "linear"
@@ -246,7 +531,8 @@ cnetplot.compareClusterResult <- function(x,
                                   fixed = TRUE), y_union$Description)
     n <- length(geneSets)
     g <- list2graph(geneSets)
-    edge_layer <- geom_edge(alpha=.8, colour='darkgrey')
+    edge_layer <- geom_edge(aes_(alpha = ~I(alpha)), colour='darkgrey',
+            show.legend = FALSE)
     if(is.null(dim(y)) | nrow(y) == 1) {
         V(g)$size <- 1
         V(g)$size[1] <- 3
@@ -264,20 +550,35 @@ cnetplot.compareClusterResult <- function(x,
             scale_size(range = range_gene_size * cex_gene) +
             labs(title= title) +
             theme(legend.position="none")    
+ 
+        # alpha_node <- rep(1, nrow(p$data))
+        # if (!is.null(hilight_category) && length(hilight_category) > 0) {
+        #     alpha_node <- rep(alpha_nohilight, nrow(p$data)) 
+        #     hilight_node <- c(hilight_category, hilight_gene)
+        #     alpha_node[match(hilight_node, p$data$name)] <- alpha_hilight
+        # }
+        # p$data$alpha <- alpha_node
+        p <- node_add_alpha(p, hilight_category, hilight_gene, alpha_nohilight, alpha_hilight)
         p <- add_node_label(p = p, data = p$data[-c(1:n),], label_size_node = label_size_gene,
             cex_label_node = cex_label_gene, shadowtext = shadowtext_gene)
         p <- add_node_label(p = p, data = p$data[1:n,], label_size_node = label_size_category,
             cex_label_node = cex_label_category, shadowtext = shadowtext_category)   
-            # geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
-            #     size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
-            # geom_node_text(aes_(label=~name), data = p$data[1:n,],
-            #     size = label_category * cex_label_category, bg.color = "white", repel=TRUE)
         return(p)
     }
 
     if(is.null(dim(y_union)) | nrow(y_union) == 1) {
         p <- ggraph(g, "tree") + edge_layer
     } else {
+        if (!is.null(hilight_category) && length(hilight_category) > 0) {
+            edges <- attr(E(g), "vnames")
+            E(g)$alpha <- rep(alpha_nohilight, length(E(g)))
+            hilight_edge <- grep(paste(hilight_category, collapse = "|"), edges)
+            hilight_gene <- edges[hilight_edge]
+            hilight_gene <- gsub(".*\\|", "", hilight_gene)
+            E(g)$alpha[hilight_edge] <- min(0.8, alpha_hilight)
+        } else {
+            E(g)$alpha <- rep(0.8, length(E(g)))
+        }
         p <- ggraph(g, layout=layout, circular=circular) + edge_layer
     }
 
@@ -341,26 +642,27 @@ cnetplot.compareClusterResult <- function(x,
                     # data=ID_Cluster_mat2,
                     # cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
                     # color=NA) +
-            p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+
+            p <- node_add_alpha(p, hilight_category, hilight_gene, alpha_nohilight, alpha_hilight)
+            ID_Cluster_mat2$alpha <-  p$data$alpha       
+            p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius,alpha=~I(alpha)),
                     data=ID_Cluster_mat2[1:n, ],
-                    cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)], color=NA) +
+                    cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-4)], 
+                    color=NA) +
                 geom_scatterpie_legend(ID_Cluster_mat2$radius[1:n],
                     x=x_loc, y=y_loc + 3, n = legend_n, labeller=function(x) round(x^2 * sum_yunion / cex_category))  +
-                geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+                geom_scatterpie(aes_(x=~x,y=~y,r=~radius,alpha=~I(alpha)),
                     data=ID_Cluster_mat2[-(1:n), ],
-                    cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
-                    color=NA, show.legend = FALSE) +
+                    cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-4)],
+                    color=NA,
+                    show.legend = FALSE) +
                 coord_equal()+
                 geom_scatterpie_legend(ID_Cluster_mat2$radius[(n+1):nrow(ID_Cluster_mat2)],
                     x=x_loc, y=y_loc, n = legend_n,
-                    labeller=function(x) round(x*2/(min(sizee))/sqrt(cex_gene),3)) +
+                    labeller=function(x) round(x*2/(min(sizee))/sqrt(cex_gene),1)) +
                 ggplot2::annotate("text", x = x_loc + 3, y = y_loc, label = "log2FC")  +
                 ggplot2::annotate("text", x = x_loc + 3, y = y_loc + 3, label = "gene number")
 
-            # p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
-            #         size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
-            #     geom_node_text(aes_(label=~name), data = p$data[1:n,],
-            #         size = label_category * cex_label_category, bg.color = "white", repel=TRUE) + 
             p <- add_node_label(p = p, data = p$data[-c(1:n),], label_size_node = label_size_gene,
                 cex_label_node = cex_label_gene, shadowtext = shadowtext_gene)
             p <- add_node_label(p = p, data = p$data[1:n,], label_size_node = label_size_category,
@@ -372,23 +674,21 @@ cnetplot.compareClusterResult <- function(x,
         # p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius), data=ID_Cluster_mat2,
                 # cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
                 # color=NA) +
-           # coord_equal()
-        p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+           # coord_equal()   
+        p <- node_add_alpha(p, hilight_category, hilight_gene, alpha_nohilight, alpha_hilight) 
+        ID_Cluster_mat2$alpha <-  p$data$alpha  
+        p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius,alpha=~I(alpha)),
                 data=ID_Cluster_mat2[1:n, ],
-                cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)], color=NA) +
-            geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+                cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-4)], color=NA) +
+            geom_scatterpie(aes_(x=~x,y=~y,r=~radius,alpha=~I(alpha)),
                 data=ID_Cluster_mat2[-(1:n), ],
-                cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
+                cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-4)],
                 color=NA, show.legend = FALSE) +
             coord_equal() +
             geom_scatterpie_legend(ID_Cluster_mat2$radius[1:n],
                     x=x_loc, y=y_loc, n = legend_n, labeller=function(x) round(x^2 * sum_yunion / cex_category)) +
             ggplot2::annotate("text", x = x_loc + 3, y = y_loc, label = "gene number")
         ## add node label
-        # p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
-        #         size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
-        #     geom_node_text(aes_(label=~name), data = p$data[1:n,],
-        #         size = label_category * cex_label_category, bg.color = "white", repel=TRUE) + 
         p <- add_node_label(p = p, data = p$data[-c(1:n),], label_size_node = label_size_gene,
             cex_label_node = cex_label_gene, shadowtext = shadowtext_gene)
         p <- add_node_label(p = p, data = p$data[1:n,], label_size_node = label_size_category,
@@ -399,26 +699,11 @@ cnetplot.compareClusterResult <- function(x,
     title <- colnames(ID_Cluster_mat2)[1]
     V(g)$size <- ID_Cluster_mat2$radius
     V(g)$color <- "#B3B3B3"
-    V(g)$color[1:n] <- "#E5C494"
-
-    ggraph(g, layout=layout, circular=circular) + 
-        edge_layer +
-        geom_node_point(aes_(color=~I(color), size=~size),
-            data = p$data[1:n, ]) +
-        scale_size(range = range_category_size * cex_category) +
-        ggnewscale::new_scale("size") +
-        geom_node_point(aes_(color=~I(color), size=~size),
-            data = p$data[-(1:n), ], show.legend = FALSE) +
-        scale_size(range = range_gene_size * cex_gene) +
-        labs(title= title)
-        # geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
-        #     size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
-        # geom_node_text(aes_(label=~name), data = p$data[1:n,],
-        #     size = label_category * cex_label_category, bg.color = "white", repel=TRUE) + 
-        p <- add_node_label(p = p, data = p$data[-c(1:n),], label_size_node = label_size_gene,
-            cex_label_node = cex_label_gene, shadowtext = shadowtext_gene)
-        p <- add_node_label(p = p, data = p$data[1:n,], label_size_node = label_size_category,
-            cex_label_node = cex_label_category, shadowtext = shadowtext_category)
-        p <- p + theme_void() + theme(legend.position="none")
+    V(g)$color[1:n] <- "#E5C494"  
+    p <- node_add_alpha(p, hilight_category, hilight_gene, alpha_nohilight, alpha_hilight)
+    p <- add_node_label(p = p, data = p$data[-c(1:n),], label_size_node = label_size_gene,
+        cex_label_node = cex_label_gene, shadowtext = shadowtext_gene)
+    p <- add_node_label(p = p, data = p$data[1:n,], label_size_node = label_size_category,
+        cex_label_node = cex_label_category, shadowtext = shadowtext_category)
+    p + theme_void() + theme(legend.position="none")
 }
-